sop_orchestrator/requirements.txt
# Core orchestration platform dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
sqlalchemy>=2.0.0
alembic>=1.13.0
redis>=5.0.0
pydantic>=2.5.0

# AI & Browser automation
browser-use>=1.0.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Web interface
jinja2>=3.1.0
python-multipart>=0.0.6

# Visual monitoring and image processing
Pillow>=10.0.0

# Monitoring & logging
structlog>=23.1.0
prometheus-client>=0.19.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# Development & testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
ruff>=0.1.0

# Optional: Local LLM support
# ollama>=0.1.0
# transformers>=4.35.0

# Core dependencies
structlog>=23.1.0
asyncio-mqtt>=0.13.0
websockets>=11.0.3
aiofiles>=23.1.0

# Web interface
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2

# Visual monitoring
Pillow>=10.0.0

# Browser automation (install browser-use separately)
# browser-use  # Install from local directory or pip

# Optional: AI model integrations
# anthropic>=0.7.0
# openai>=1.0.0 